+++ client/Packages/com.beamable.server/SharedRuntime/Api/Chat/IMicroserviceChatApi.cs
@@ -1,0 +1,13 @@
+using Beamable.Common;
+using Beamable.Common.Api;
+using Beamable.Experimental.Api.Chat;
+using System.Collections.Generic;
+
+namespace Beamable.Server.Api.Chat;
+
+/// <summary>
+/// The chat API for Microservice usage
+/// </summary>
+public interface IMicroserviceChatApi : IChatApi
+{
+}