+++ client/Packages/com.beamable/CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - List of available to create `ContentTypes` in `Content Manager` contextual menu is now ordered alphabetically
 - The Beamable host URL is no longer sourced from `config-defaults.txt`. Instead, it comes from the `BeamableEnvironment` class. 
 - Changed `PackageVersion` to accept "preview" prefix strings instead of requiring a direct match of the string "preview"
+- Moved `JsonSerializable` to Beamable.Common assembly
+- Moved some parts of the `ChatService` to Beamable.Common assembly
+- Changed namespace of `Beamable.Pooling.ClassPool` to `Beamable.Common.Pooling.ClassPool`
 
 ### Fixed
 - Beamable button in Unity toolbar should be in correct position for production packages